#875 BranchCity calls ShowMarking after InitializeAfterDrawn. The problem is that the EdgeMeshScheduler must be initialized when the underlying graph has been changed. It is an observer to a particular graph. When we move from one graph to the next one, it needs to register itself by the new graph. Otherwise, new edges remain LineRenderers. We need them to be meshes, however, because otherwise the highlight effects wouldn't work and they could neither be selected.
Added a few more FIXMEs/TODOs. We cannot update the edge layout while
moving the nodes, because the nodes have been unparented. The hierarchy
is, however, necessary for edge bundling layout. We could fade
out the edges, then move the nodes, and then fade in the edges with the
new layout again.

Question: Do we really need to unparent the nodes before moving them?