Change by Unity.
The overrideShaderVariantLimit property in Unity's ShaderGraphSettings.asset is a value that, when set to a number greater than zero, overrides the project-wide maximum limit for the number of shader variants that a single Shader Graph can generate.

It is part of Unity's mechanism to prevent shaders from becoming overly complex, which can lead to extremely long build times and massive build sizes.