Fixing premature execution of fetching multiplayer data from backend.
Under some circumstances ActionNetwork.Start() could be called while
the network is not yet up and running. The IsServer, IsHost and IsClient
are all false then; as a consequence RequestSynchronizationServerRpc()
was called and failed.

As @matthisw pointed out, the Unity documentation recommends to call
such code in OnNetworkSpawn() rather than Start() methods.

All credits of this change go to @matthisw.