#887 Removed redundant code to retrieve an AbstractSEECity from objects.
There were many places where SceneQueries.GetCodeCity was called to
retrieve the component AbstractSEECity. This functionality is now
available in SceneQueries.City. The redundant code was replaced by a
call to that.

Note: There is another similar method in GameObjectExtensions.ContainingCity.
We will consider to replace SceneQueries.City by that one. It is nicer
to use an extension.