#887 The label is prepared only when we have a valid NodeRef.
A valid NodeRef is one whose Value differs from null.
For the BranchCity (if created in the Editor) it can happen that we
create a NodeOperator before the graph is actually deserialized and all
NodeRefs properly set, that is, when the NodeRef is not yet valid.
In that case, we postpone the label preparation until it becomes
available. The label preparation needs to know the node to retrieve
the name of the node to be shown.